+++ packages/pagination/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@alfalab/core-components-table",
+    "name": "@alfalab/core-components-pagination",
     "version": "1.0.0",
     "description": "",
     "keywords": [],
@@ -18,8 +18,7 @@
     },
     "dependencies": {
         "@alfalab/core-components-button": "^5.0.0",
-        "@alfalab/core-components-pagination": "^1.0.0",
-        "@alfalab/core-components-select": "^10.2.0",
+        "@alfalab/core-components-tag": "^4.0.0",
         "classnames": "^2.2.6"
     }
 }
