+++ packages/pagination/src/index.module.css
@@ -1,0 +1,28 @@
+@import '../../themes/src/default.css';
+
+.component {
+    display: flex;
+}
+
+.tag.tag {
+    flex-shrink: 0;
+    margin-right: var(--gap-xs);
+    padding: 0 var(--gap-xs);
+    min-width: 32px;
+
+    &:last-child {
+        margin-right: 0;
+    }
+
+    &.tagActive {
+        cursor: default;
+    }
+}
+
+.menu {
+    margin: var(--gap-2xs) 0;
+}
+
+.option {
+    min-width: 208px;
+}
